requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai>=1.50.0
tiktoken==0.7.0
pypdf==4.2.0
python-pptx==0.6.23
python-docx==1.1.0
Pillow==10.4.0
pytesseract==0.3.10
pandas==2.2.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
rank-bm25==0.2.2
tavily-python==0.3.5
